Some bug fixes and new features
FEATURE: It's now possible to search for iPhone, iPad and Mac apps;
FEATURE: It's now possible to return more than on result, for
situtation when the first result isn't the most accurate;
FEATURE: You can now choose if you want simply the link of a markdown
style link with the correct app name as well as affiliate link.

BUG: URL wasn't percent-encoded twice as instructed by Apple
documentation.